requirements.txt
# Core runtime
fastapi>=0.110,<1
uvicorn>=0.29,<1
pydantic>=2,<3

# Persistence
sqlalchemy>=2,<3
pymysql>=1,<2

# Graph integration
neo4j>=5,<6

# Testing
pytest>=8,<9
httpx>=0.27,<1

# CDC stream ingestion
kafka-python>=2,<3
